Basic component props adjust
